[feat](#11) ProgressReader를 통해 SRP 원칙 준수 + progress 관련 클래스 위치 변경
