+++ src/main/java/com/poweranger/hai_duo/progress/api/controller/UserProgressController.java
@@ -1,8 +1,8 @@
-package com.poweranger.hai_duo.quiz.api.controller;
+package com.poweranger.hai_duo.progress.api.controller;
 
 import com.poweranger.hai_duo.progress.api.dto.ChapterResponseDto;
 import com.poweranger.hai_duo.progress.api.dto.LevelResponseDto;
-import com.poweranger.hai_duo.quiz.api.dto.ProgressResponseDto;
+import com.poweranger.hai_duo.progress.api.dto.ProgressResponseDto;
 import com.poweranger.hai_duo.progress.api.dto.StageResponseDto;
 import com.poweranger.hai_duo.user.application.service.UserProgressService;
 import lombok.RequiredArgsConstructor;
