+++ src/main/java/com/poweranger/hai_duo/progress/api/dto/ProgressResponseDto.java
@@ -1,4 +1,4 @@
-package com.poweranger.hai_duo.quiz.api.dto;
+package com.poweranger.hai_duo.progress.api.dto;
 
 import lombok.Builder;
